Improvements to downloading and game management
Changes have been made to the `DownloadItem.cs` file, adding a new `System.Diagnostics` namespace and improving the `Progress` property logic and handling of download cancellation. Changed the way the file is opened to better handle errors and added closing the file stream in the `finally` block.

In the `DownloadManager.cs` file, the signature of the `StartDownloadAsync` method was changed to accept an additional `gameLogo` parameter, and unused code was removed.

In `GameBaseViewModel.cs`, added support for situations where `gameSymbol` can be `null`, and a new `DownloadGame` method for adding download tasks. The logic for loading game logos has also been improved.

In `GameBasePage.xaml.cs`, the way the download method is called has been updated to use the new `DownloadGame` method.

In `ShellPage.xaml.cs`, the signature of the `AddDownload` method has been changed to accept an additional `gameLogo` parameter, and a new `RemoveDownload` method has been added to remove download tasks. Also added support for the `OnLoaded` event to update the title bar.