Refactoring the game logic and updates in the UI
Removed the line responsible for deleting the zip file in `LauncherStatusHelper.cs`, which may affect memory management. Added a new `PlayGame()` method in `GameBaseViewModel.cs`, which starts the game and updates the game time in the database. Changed button click events in `GameBasePage.xaml` to call new methods in `GameBaseViewModel`. Updated the logic in `GameBasePage.xaml.cs` to correctly handle the game ID and removed some methods related to updates. Made changes to the `PlayButton_Click` method to take advantage of the new logic. Corrected the game removal logic in `UninstallButton_Click`.