Improved installation location and file management
Changed the storage of installation locations in `ChooseInstallLocationDialog.xaml.cs` from `SelectedPath` to `SelectedFolder` (StorageFolder) and added support for `FolderPicker`.

In `DownloadItem.cs`, the `DestinationPath` type was updated to `StorageFolder`, which improves folder management.

In `DownloadManager.cs` and `GameBaseViewModel.cs`, updated methods to use `StorageFolder` and added support for `StorageApplicationPermissions.FutureAccessList`.

Adjusted logic in `ShellPage.xaml.cs` to handle ZIP files and downloads, improving the performance and security of file operations.